requirements.txt
Pillow~=8.3.1
opencv-python~=4.5.3.56
pip~=20.0.2
numpy~=1.21.1
setuptools~=44.0.0
certifi~=2021.5.30
urllib3~=1.26.6
idna~=3.2
selenium~=3.141.0
pytesseract~=0.3.8
requests~=2.26.0
python-dotenv~=0.19.0